# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Order Hub API — multi-platform inventory and order management
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /purchases/product-names:
    get:
      operationId: listPurchaseProductNames
      tags: [purchases]
      summary: Distinct product names used across past purchases (for autocomplete)
      parameters:
        - in: query
          name: search
          description: Case-insensitive substring filter on product name
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
      responses:
        "200":
          description: Suggested product names sorted by most recently used
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [productName]
                      properties:
                        productName:
                          type: string
                        lastUsed:
                          type: string
                          nullable: true
                          format: date-time
  /purchases:
    get:
      operationId: listPurchases
      tags: [purchases]
      summary: Paginated list of manually-recorded purchases (active company or all)
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 1
            default: 1
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 25
        - in: query
          name: search
          schema:
            type: string
        - in: query
          name: sortBy
          schema:
            type: string
            enum: [purchaseDate, total, quantity, rate, productName]
            default: purchaseDate
        - in: query
          name: sortOrder
          schema:
            type: string
            enum: [asc, desc]
            default: desc
        - in: query
          name: purchaseDateFrom
          description: Inclusive lower bound for purchaseDate (ISO 8601 date or datetime)
          schema:
            type: string
        - in: query
          name: purchaseDateTo
          description: Inclusive upper bound for purchaseDate (ISO 8601 date or datetime)
          schema:
            type: string
        - in: query
          name: allCompanies
          description: When true, returns purchases across every company instead of just the active brand
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Purchase list with totals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PurchaseList"
    post:
      operationId: createPurchase
      tags: [purchases]
      summary: Record a new purchase entry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /purchases/{id}:
    put:
      operationId: updatePurchase
      tags: [purchases]
      summary: Update a purchase entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PurchaseInput"
      responses:
        "200":
          description: Updated purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Purchase"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deletePurchase
      tags: [purchases]
      summary: Delete a purchase entry
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin only)
      responses:
        "200":
          description: User list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AppUser"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/{id}:
    patch:
      operationId: updateUserRole
      tags: [users]
      summary: Update a user's role (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserRolePayload"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppUser"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteUser
      tags: [users]
      summary: Remove a user from the workspace (admin only)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Brand:
      type: string
      description: |
        Active brand workspace identifier (a company `slug`). Sent on every
        authenticated request via the `X-Brand` header. Companies are managed
        dynamically via the `/companies` endpoints.
...
    Role:
      type: string
      enum: [admin, staff]
...
    AppUser:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        email:
          type: string
        fullName:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        role:
          $ref: "#/components/schemas/Role"
        createdAt:
          type: string
          format: date-time
      required: [id, clerkUserId, email, role, createdAt]
    UpdateUserRolePayload:
      type: object
      properties:
        role:
          $ref: "#/components/schemas/Role"
      required: [role]
...
    Purchase:
      type: object
      properties:
        id:
          type: integer
        brand:
          $ref: "#/components/schemas/Brand"
        productName:
          type: string
        sku:
          type: string
        quantity:
          type: integer
        rate:
          type: number
          description: Per-unit purchase rate
        total:
          type: number
          description: rate * quantity
        vendor:
          type: ["string", "null"]
        partyName:
          type: ["string", "null"]
        category:
          type: ["string", "null"]
        purchaseDate:
          type: string
          format: date-time
        notes:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        amazonSku:
          type: ["string", "null"]
        amazonSellingPrice:
          type: ["number", "null"]
        amazonGstRate:
          type: ["number", "null"]
          description: GST rate (%) for Amazon
        flipkartSku:
          type: ["string", "null"]
        flipkartSellingPrice:
          type: ["number", "null"]
        flipkartGstRate:
          type: ["number", "null"]
          description: GST rate (%) for Flipkart
        meeshoSku:
          type: ["string", "null"]
        meeshoSellingPrice:
          type: ["number", "null"]
        meeshoGstRate:
          type: ["number", "null"]
          description: GST rate (%) for Meesho
        createdAt:
          type: string
          format: date-time
      required: [id, brand, productName, sku, quantity, rate, total, vendor, partyName, category, purchaseDate, notes, imageUrl, amazonSku, amazonSellingPrice, amazonGstRate, flipkartSku, flipkartSellingPrice, flipkartGstRate, meeshoSku, meeshoSellingPrice, meeshoGstRate, createdAt]
    PurchaseInput:
      type: object
      properties:
        productName:
          type: string
          minLength: 1
        sku:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 1
        rate:
          type: number
          minimum: 0
        vendor:
          type: string
        purchaseDate:
          type: string
          format: date-time
        notes:
          type: string
        imageUrl:
          type: string
        partyName:
          type: string
        category:
          type: string
        amazonSku:
          type: string
        amazonSellingPrice:
          type: number
          minimum: 0
        amazonGstRate:
          type: number
          minimum: 0
          maximum: 100
        flipkartSku:
          type: string
        flipkartSellingPrice:
          type: number
          minimum: 0
        flipkartGstRate:
          type: number
          minimum: 0
          maximum: 100
        meeshoSku:
          type: string
        meeshoSellingPrice:
          type: number
          minimum: 0
        meeshoGstRate:
          type: number
          minimum: 0
          maximum: 100
      required: [productName, sku, quantity, rate, purchaseDate]
    PurchaseTotals:
      type: object
      properties:
        totalCost:
          type: number
        totalUnits:
          type: integer
        totalEntries:
          type: integer
      required: [totalCost, totalUnits, totalEntries]
    PurchaseList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Purchase"
        totals:
          $ref: "#/components/schemas/PurchaseTotals"
        page:
          type: integer
        pageSize:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
      required: [items, totals, page, pageSize, total, totalPages]